requirements.txt
reval
pandas
numpy
flask
matplotlib
requests
Jinja2
openai
python-dotenv